Handle delete-modify conflicts in staging view
Delete-modify conflicts could not be resolved in the staging view by
removing the file. If a conflicting file has been removed from the
working tree, make staging it in the staging view remove it from the
index.

Note that if the preference for "Automatically stage files being
deleted" is switched *on*, then deleting the conflicting file in the
working tree would remove it from the index already. But with that
preference *off* the user had no real way to resolve such conflicts.
(Sometimes, this could be worked around by first staging the conflict,
then unstaging, then deleting the file, and then staging the deletion.)

The solution in this change works, but still leaves some non-working
commands in the context menu. There's no delete command in the context
menu, and the "replace with" commands (ours/theirs/HEAD/index) don't
do anything for the side that deleted the file.

Fixing that is left for a future enhancement.

Bug: 546869
Change-Id: I9594587f9445acf0a9972b01933bc5d453168f65
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>