Enable the "Switch To..." menu for submodule folders
Commits 1f69fad3 and e4dc3eaf from the multi-repository branch
switch had the effect that the "Switch To..." menu didn't work
anymore for submodule folders in the package or project explorer.

Add a new operation SelectionUtils.getAllRepositories() and use it
in SwitchToMenu. Changing the existing SelectionUtils.getRepositories()
to also return repositories for folders was deemed too risky since
that would have changed the behavior of the multi-repository pull,
which has always worked only for folders. Enabling that to also handle
submodule folders might be an idea for a future commit, but might
need more tests.

Add a new test. The test only verifies that the menu has content by
trying to invoke "New Branch..." and then closing the resulting
"Create Branch" dialog.

Interestingly this test succeeds even without the fix, which matches
the experience described in the bug report: the problem could only
be observed in official packages, but not in my development setup.

Change verified manually by actually building EGit and installing it
into standard Eclipse packages.

Bug: 548158
Change-Id: Ic67578b46a89687a6bcfe73ef3d2d3d71d2a39e0
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>