Show standard message box on paste to repo view error
When pasting invalid content to the git repositories view, it shows the
"Problem occurred" message box due to using the status handler of
eclipse. Since this is bad user input, we should only show a message
box, and not handle this situation like an internal error.

Changed message texts to be sentence style.

Bug: 549618
Change-Id: Id6e2b4facff5deb945275e89a85d74442aa1462a
Signed-off-by: Michael Keppler <Michael.Keppler@gmx.de>