[history] Don't allow empty ref filters
Give the text cell editor in the ref filter dialog a validator and
add a label where we can display a validation error message. Use an
ICellEditorListener to update the message and the buttons. Disable
the OK button when the cell editor value is invalid.

Also prevent pasting multi-line text into the single-line control
of the cell editor. (See bug 273470.)

Bug: 553070
Change-Id: I660428ad1306d8fdfea1918a13ba040cc2b68f27
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>