Branch selection dialog improvements
This dialog can be invoked via the BranchActionHandler, which is
a default handler. Make sure the dialogs say in which repository
they operate; it may not be obvious if invoked via Quick Access.
Pre-select the current branch, if any, in the dialog, so that when
clicking "New branch..." there is a default base set.

Also give the BranchActionHandler a description.

Bug: 553223
Change-Id: I0aaf4a6788382201310dbcd1ff4f154577720db6
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>