Auto-sharing must not connect root or home directories
When there are multiple mappings found for a project, EGit used
to add the last one (the topmost one) found to the configured
repositories. This is to avoid adding submodules as configured
repositories.

But just taking the topmost repository found is not good; it could
lead to adding ~/.git (if that exists) to the configured repositories.
A repo at ~/.git is likely to have to have a huge working tree with
millions of files, and it's almost always a user error (for instance,
typing "git init" by mistake in the home directory and then forgetting
about it).

Instead walk up the list of nested repositories found. If an already
configured repository is found, stop. Otherwise, add the topmost one
that has still a "valid" path, i.e., that is neither at ~ or above nor
at root.

Bug: 553201
Change-Id: I035f2a24844e7c253401186243db9e3716c41c48
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>