[repo view] Work-around for SWT regression on GTK
Remove the focus cell manager. It is needed only for keyboard
activation of inline cell editing, but we don't use that anyway
since we always go through the rename command, which is a
programmatic editor activation.

The focus manager ensures the focus cell is visible by calling
Tree.showItem(), which is specified to do nothing if the cell is
already visible. But since the changes made in the GTK version of
SWT for bug 552096, showItem() _always_ scrolls the item to the
center of the tree, which is quite annoying and makes double-clicking
an item impossible unless it already is in the center of the view.

As a work-around, remove that unnecessary focus manager.

Bug: 558496
Change-Id: I69f5c4274350a909dd0288a0f8650346ba8da175
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>