Enable and execute cleanup "Push down negations"
Push down negations to make reading of complex expressions with
negations easier.

Also enable the cleanup as a save action. There is a bug in the cleanup
(not adding braces around instanceof expressions), but whenever that
occurs, the developer can just add the braces manually, and the bug will
not occur again for that location. Therefore that bug should not stop us
from using the cleanup.

Change-Id: I1a5026304b080b0ccebe06e6829857b784541f99
Signed-off-by: Michael Keppler <Michael.Keppler@gmx.de>