[diff editor] Enable "Save As..." on unified diffs
Contrary to the normal "Save As..." implementation, do _not_ reload
the editor with the newly created FileEditorInput. Just save the
document to a file; leave opening the file in an editor to the user.

To make the "Save As..." handler update its enablement properly in all
cases when the nested editor pages switch force a re-evaluation of the
active part context variable on page changes.

Change-Id: Ia3e8bc1cd81e40fbf7511a32fee1967104b67098
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>