[unified diff] Correct hyperlinks and outline context menu
Use "previous version" only if the diff is indeed against the first
parent. Otherwise use "base version".

Also show the commit id of the version, and add the short commit
message as a tooltip on the context menu entries. (Shows only on
newer Eclipse platforms; c.f. bug 473959.)

Try to show the link for opening the workspace version also in the
hyperlinks for the old version, if possible. Because for a "copy"
diff it would be very strange to have a link to show the new file
on an "old version" link, try to open the old file name if it still
exists. To get still a somewhat consistent behavior, don't show the
link on the old version if the file does not exist. Showing a link
to open the new version there would give very confusing behavior.
Also add the file name to the link text.

As a result, the "Open Working Tree Version" link for the "old
version" is

* ADD, DELETE, and RENAME diff: absent
* MODIFY diff: present, same file as "new version"
* COPY diff: present, old file name

Bug: 561012
Change-Id: I67177e4f71a4f295bd26a297ebfbecc7abdaac43
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>