Remove logging that server is a Gerrit server
If I clone a repo from Gerrit, my Error log view gets focus and shows me
the message that the server was detected to be a Gerrit server. I think
this is more disturbing (as it lands in the Error Log view) than helpful
and suggest to remove this logging.

Change-Id: I44ab51c1cee50ee8385dfd8c16c8dfc6ed9f9b6e
Signed-off-by: Lars Vogel <Lars.Vogel@vogella.com>