Use TREE_TABLE_FONT if available
EGit has custom font preferences for ignored resources or resources
with uncommitted changes. Eclipse 4.17 has a new font preference
allowing the user to change the font in trees and tables. When he
does change that font, the EGit-styled resources do not follow suit
because the EGit font definitions have no default, and we can't set
a default since the new TREE_TABLE_FONT doesn't exist on older
Eclipses, where we would thus get exceptions from the framework.

It's also not possible with the current build setup to move these font
definitions to fragments, one for Eclipse < 4.17 and one for >= 4.17.
At the very least this would require assembling the p2 repository with
an Eclipse 4.17 target platform, even if the bundles had been compiled
before against an Eclipse 4.6 target.

Instead just do the work explicitly when decorating resources and
check for the new font if EGit's own preference has not been set (in
which case it would override the default TREE_TABLE_FONT anyway).

Bug: 564970
Change-Id: I9a9f160235b95845ededdaf023bfe1b2145661a1
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>