DiffEditorOutlinePage: swap icons for "Toggle Presentation" button
This makes the button icon and tooltip describe what will happen
when the user clicks the button. Previous code showed a list icon
that when clicked would change the presentation to a compact tree,
or a folder icon while clicking it would change to a flat list.
That's confusing.

Newly added tooltips clarify the meaning of the button.

Change-Id: I90bb1ef316502dac9e9ada6309a4c51173ca18fd
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>