Allow to open a commit from the properties view
When the properties view shows a commit (currently only as target of
a tag), provide a way to open that commit in the commit viewer, or to
show it in the history view, both directly from the properties view.

Since the context menu on the viewer used in the PropertySheetPage is
not accessible and not extendable either, use a CellEditor that shows
two buttons to perform the actions. (A ToolBar cannot be used since
we couldn't capture the Esc-key.)

Change-Id: I75827ee7217a186d52cbc22611ce958e8fe85f3b
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>