Manually fix license status of dependencies
For the following p2 dependencies the license tool detected status for
the binary artefact but not for the corresponding source artefact:
- org.apache.commons.codec/1.14.0.v20200818-1422
- org.apache.commons.compress/1.19.0.v20200106-2343
- org.apache.sshd.osgi/2.6.0.v20210201-2003
- org.apache.sshd.sftp/2.6.0.v20210201-2003
- org.bouncycastle.bcpg/1.65.0.v20200527-1955
- org.bouncycastle.bcpkix/1.65.0.v20200527-1955
- org.bouncycastle.bcprov/1.65.1.v20200529-1514
Copy the auto-detected status from the binary artefact to the
corresponding source artefact.

For
- org.apache.ant/1.10.9.v20201106-1946
the license tool detected license status for the source artefact but not
for the binary artefact. Copy the auto-detected status from the source
artefact to the corresponding source artefact.

For
- org.slf4j.binding.log4j12.source/1.7.30.v20201108-2042
the license tool did not find the corresponding CQ probably since it
failed to find the Maven artefact name from the bundle name of the
p2 dependency. Copy the auto-detected license status from the
corresponding entry in JGit where the same dependency is given as a
Maven dependency to
maven/mavencentral/org.slf4j/slf4j-log4j12/1.7.30, MIT, approved, CQ7665

Change-Id: I85024457a809e54f447abbb16fb2cb5db7bacda1