Simplify labels for "Replace With->Ours/Theirs"
While it may have been a nice touch to try to determine the last commit
that actually touched a file and show that commit message in the menu
label, there were several problems with this. The user is merging two
commits. It doesn't help him if the label shows a different commit. For
conflicts during rebasing or cherry-picking, the determination of the
last commit to have touched a file was not done. The determination can
also be rather expensive.

So simplify this and always use the two commits that the user operated
on.

Change-Id: I28a3d5a81c5ec0fa13fd69efc4b469bb342ffa73
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>